removes repeated code from logging and uses module name in getLogger
