+++ src-api/openzwave/command.py
@@ -23,20 +23,14 @@ You should have received a copy of the GNU General Public License
 along with python-openzwave. If not, see http://www.gnu.org/licenses.
 
 """
+
+import logging
+
 from openzwave.object import ZWaveNodeInterface
 from threading import Timer
 
-# Set default logging handler to avoid "No handler found" warnings.
-import logging
-try:  # Python 2.7+
-    from logging import NullHandler
-except ImportError:
-    class NullHandler(logging.Handler):
-        """NullHandler logger for python 2.6"""
-        def emit(self, record):
-            pass
-logger = logging.getLogger('openzwave')
-logger.addHandler(NullHandler())
+logger = logging.getLogger(__name__)
+
 
 class ZWaveNodeBasic(ZWaveNodeInterface):
     """
