+++ src-api/openzwave/option.py
@@ -23,7 +23,9 @@ You should have received a copy of the GNU General Public License
 along with python-openzwave. If not, see http://www.gnu.org/licenses.
 
 """
+
 import os
+import logging
 from platform import system as platform_system
 
 import libopenzwave
@@ -31,17 +33,9 @@ from libopenzwave import PyLogLevels
 from openzwave.object import ZWaveException
 from openzwave.singleton import Singleton
 
-# Set default logging handler to avoid "No handler found" warnings.
-import logging
-try:  # Python 2.7+
-    from logging import NullHandler
-except ImportError:
-    class NullHandler(logging.Handler):
-        """NullHandler logger for python 2.6"""
-        def emit(self, record):
-            pass
-logger = logging.getLogger('openzwave')
-logger.addHandler(NullHandler())
+
+logger = logging.getLogger(__name__)
+
 
 def _expand_path(path):
     if path is None:
