+++ src-api/openzwave/value.py
@@ -23,20 +23,15 @@ You should have received a copy of the GNU General Public License
 along with python-openzwave. If not, see http://www.gnu.org/licenses.
 
 """
+
+import logging
+
 from six import string_types
 from openzwave.object import ZWaveObject
 
-# Set default logging handler to avoid "No handler found" warnings.
-import logging
-try:  # Python 2.7+
-    from logging import NullHandler
-except ImportError:
-    class NullHandler(logging.Handler):
-        """NullHandler logger for python 2.6"""
-        def emit(self, record):
-            pass
+
 logger = logging.getLogger('openzwave')
-logger.addHandler(NullHandler())
+
 
 # TODO: don't report controller node as sleeping
 # TODO: allow value identification by device/index/instance
